style: refactor packages
