+++ src/main/java/com/moviereservation/movie_reservation_system/services/AuthService.java
@@ -1,0 +1,11 @@
+package com.moviereservation.movie_reservation_system.services;
+
+import com.moviereservation.movie_reservation_system.models.user.dto.AuthResponse;
+import com.moviereservation.movie_reservation_system.models.user.dto.LoginDTO;
+import com.moviereservation.movie_reservation_system.models.user.dto.RegisterDTO;
+
+public interface AuthService {
+
+    AuthResponse registerUser(RegisterDTO registerDTO);
+    AuthResponse loginUser(LoginDTO loginDTO);
+}
